step4 - 피드백 반영 - 강의레파지토리와 강의신청레파지토리 분리
